refine predicate on SubsetType to act on denoted values, separate tests, add line to docs
