+++ src/test/scala/cosas/DenotationTests.scala
@@ -8,31 +8,7 @@ object DenotationTestsContext {
   object User extends Type("User")
   type User = User.type
   object Friend extends Type("Friend")
-  case class userInfo(id: String, name: String, age: Int)
-
-  /* The NEList stuff */
-  final class WrappedList[E] extends Wrap[List[E]]("WrappedList")
-
-  class NEList[E] extends SubsetType[WrappedList[E]] {
-
-    lazy val label = "NEList"
-    def predicate(l: List[E]): Boolean = ! l.isEmpty
-
-    def apply(e: E): ValueOf[NEList[E]] = new ValueOf[NEList[E]](e :: Nil)
-  }
-
-  object NEList {
-
-    implicit def toOps[E](v: ValueOf[NEList[E]]): NEListOps[E] = new NEListOps(v.value)
-    implicit def toSSTops[E](v: NEList[E]): SubSetTypeOps[WrappedList[E], NEList[E]] = new SubSetTypeOps(v)
-  }
-
-  def NEListOf[E]: NEList[E] = new NEList()
-
-  class NEListOps[E](val l: List[E]) extends AnyVal with ValueOfSubsetTypeOps[WrappedList[E], NEList[E]] {
-
-    def ::(x: E): ValueOf[NEList[E]] = unsafeValueOf[NEList[E]](x :: l)
-  }
+  case class userInfo(id: String, name: String, age: Int)  
 }
 
 class DenotationTests extends org.scalatest.FunSuite with ScalazEquality {
@@ -91,18 +67,5 @@ class DenotationTests extends org.scalatest.FunSuite with ScalazEquality {
     val azul = Color := "blue"
 
     assert(azul.show == "(Color := blue)")
-  }
-
-  test("naive nonempty lists") {
-
-    import DenotationTestsContext._
-
-    import AnySubsetType._
-    // this is Some(...) but we don't know at runtime. What about a macro for this? For literals of course
-    val oh = NEListOf[Int](12 :: 232 :: Nil)
-
-    val nelint = NEListOf(232)
-
-    val u1 = 23 :: nelint
   }
 }
